Switch distro to a live-updateable config file
The distro used to be hardcoded in the script. Now it is encoded in a
config file that is sourced not merely at runtime (when the script is
first started up), but each time there is a new notification to send
out. As such, the distro can be updated without restarting the script.